+++ src/main/java/sws/songpin/domain/member/controller/HomeController.java
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import sws.songpin.domain.member.service.HomeService;
 import sws.songpin.domain.member.service.MemberService;
 
 @Tag(name = "Home", description = "Home 기능 관련 API입니다.")
@@ -14,12 +15,12 @@ import sws.songpin.domain.member.service.MemberService;
 @RequiredArgsConstructor
 @RequestMapping("/home")
 public class HomeController {
-    private final MemberService memberService;
+    private final HomeService homeService;
 
     @Operation(summary = "홈 페이지", description = "(1) 인사말, (2) 최근 등록된 핀 3개, (3) 최근 등록된 장소 3개")
     @GetMapping
     public ResponseEntity<?> getHome(){
-        return ResponseEntity.ok(memberService.getHome());
+        return ResponseEntity.ok(homeService.getHome());
     }
 
 }
