+++ src/main/java/sws/songpin/domain/member/service/MemberService.java
@@ -7,30 +7,18 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import sws.songpin.domain.member.dto.response.HomeResponseDto;
 import sws.songpin.domain.member.dto.response.MemberSearchResponseDto;
 import sws.songpin.domain.member.dto.response.MemberUnitDto;
 import sws.songpin.domain.member.entity.Member;
 import sws.songpin.domain.member.repository.MemberRepository;
-import sws.songpin.domain.pin.dto.response.PinBasicUnitDto;
-import sws.songpin.domain.pin.entity.Pin;
-import sws.songpin.domain.pin.repository.PinRepository;
-import sws.songpin.domain.place.dto.response.PlaceUnitDto;
-import sws.songpin.domain.place.entity.Place;
-import sws.songpin.domain.place.repository.PlaceRepository;
 import sws.songpin.global.exception.CustomException;
 import sws.songpin.global.exception.ErrorCode;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
 public class MemberService {
     private final MemberRepository memberRepository;
-    private final PinRepository pinRepository;
-    private final PlaceRepository placeRepository;
 
     // 유저 검색
     @Transactional(readOnly = true)
@@ -66,24 +54,5 @@ public class MemberService {
 
     public Member saveMember(Member member){
         return memberRepository.save(member);
-    }
-
-    @Transactional(readOnly = true)
-    public HomeResponseDto getHome() {
-        Member currentMember = getCurrentMember();
-        List<Pin> pins = pinRepository.findTop3ByOrderByPinIdDesc();
-        List<Place> places = placeRepository.findTop3ByOrderByPlaceIdDesc();
-        // pinList
-        List<PinBasicUnitDto> pinList = pins.stream()
-                .map(pin -> PinBasicUnitDto.from(pin, pin.getMember().equals(currentMember)))
-                .collect(Collectors.toList());
-        // placeList
-        List<PlaceUnitDto> placeList = places.stream()
-                .map(place -> {
-                    int placePinCount = place.getPins().size();
-                    return PlaceUnitDto.from(place, placePinCount);
-                })
-                .collect(Collectors.toList());
-        return HomeResponseDto.from(currentMember, pinList, placeList);
     }
 }
